add export selected, change names of export files

+++ frontend/src/app/modules/admin/admin-dashboard/admin-dashboard.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
 import { MatFormFieldModule } from '@angular/material/form-field';
-import { FormBuilder, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatInputModule } from '@angular/material/input';
 import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
@@ -12,26 +12,7 @@ import { MatRadioModule } from '@angular/material/radio';
 import { UserUpdateComponent } from '../user-update/user-update.component';
 import { MatDialog } from '@angular/material/dialog';
 import { UserService } from 'app/modules/services/user.service';
-
-export interface PeriodicElement {
-  user_id: string;
-  username: string;
-  email: number;
-  disabled: Date;
-}
-
-const ELEMENT_DATA: PeriodicElement[] = [
-  { user_id: '1', username: 'Hydrogen', email: 1.0079, disabled: new Date() },
-  { user_id: '2', username: 'Helium', email: 4.0026, disabled: new Date() },
-  { user_id: '3', username: 'Lithium', email: 6.941, disabled: new Date() },
-  { user_id: '4', username: 'Beryllium', email: 9.0122, disabled: new Date() },
-  { user_id: '5', username: 'Boron', email: 10.811, disabled: new Date() },
-  // { user_id: '6', username: 'Carbon', email: 12.0107, disabled: new Date() },
-  // { user_id: '7', username: 'Nitrogen', email: 14.0067, disabled: new Date() },
-  // { user_id: '8', username: 'Oxygen', email: 15.9994, disabled: new Date() },
-  // { user_id: '9', username: 'Fluorine', email: 18.9984, disabled: new Date() },
-  // { user_id: '10', username: 'Neon', email: 20.1797, disabled: new Date() },
-];
+import { SnackBarService } from 'app/modules/services/snack-bar.service';
 
 @Component({
   selector: 'app-admin-dashboard',
@@ -49,9 +30,9 @@ export class AdminDashboardComponent implements OnInit {
   isLoading : boolean
 
   constructor(
-    private fb: FormBuilder,
     public dialog: MatDialog,
-    private userService: UserService
+    private userService: UserService,
+    private snackBarService: SnackBarService
   ) {
   }
   ngOnInit(): void {
@@ -83,22 +64,19 @@ export class AdminDashboardComponent implements OnInit {
     this.searchTerm = '';
   }
 
-  submit() {
-
-  }
-
-  clearUserForm() {
-
-  }
-
   updateUser(userId: any) {
     this.openDialog(userId)
   }
 
   deleteUser(userId: any) {
-    this.userService.deleteById(userId).subscribe((res: any) => {
-      this.loadUsers()
-    })
+    let confirmation = confirm("Are you sure you want to delete this user?");
+
+    if (confirmation) {
+      this.userService.deleteById(userId).subscribe((res: any) => {
+        this.snackBarService.showSnackbar(res)
+        this.loadUsers()
+      })
+    }
   }
 
   openDialog(userId?: number) {
